platform-overview: rewrite spell-checking page
There is now the gspell library for GTK+ text widgets integration.
gspell can do everything that GtkSpell can do, and much more. So
GtkSpell is no longer recommended, see:
https://wiki.gnome.org/Projects/gspell/FAQ

gspell has a spell checker dialog, so it is not only for "real-time
spell checking". So, rename the description of the page to be more
general.

Replace Empathy by gedit, Empathy is no longer well maintained so it is
not a good example.

https://bugzilla.gnome.org/show_bug.cgi?id=781265