hig: apps can run without windows open in some cases
In some cases it is appropriate for an app to run without any
visible windows (examples include a chat or mail app that's
checking your messages in the background). We should have more
complete documentation on this, but in the mean time it's best
not to have incorrect instructions in the hig.